fix: dev: Reduce the size of __init__.py to 717 lines of code.
